+++ src/processmanager.h
@@ -9,6 +9,7 @@ class ProcessManager : public QObject {
   Q_OBJECT
   Q_PROPERTY(QString output READ output NOTIFY outputChanged)
   Q_PROPERTY(bool isRunning READ isRunning NOTIFY isRunningChanged)
+  Q_PROPERTY(bool isPaused READ isPaused NOTIFY isPausedChanged)
 
 public:
   explicit ProcessManager(QObject *parent = nullptr);
@@ -16,6 +17,7 @@ public:
 
   QString output() const { return m_output; }
   bool isRunning() const { return m_isRunning; }
+  bool isPaused() const { return m_isPaused; }
 
   Q_INVOKABLE void runCommand(const QString &program,
                               const QStringList &arguments = QStringList());
@@ -28,10 +30,13 @@ public:
                                  const QStringList &arguments = QStringList(),
                                  const QString &workingDirectory = QString());
   Q_INVOKABLE void clearOutput();
+  Q_INVOKABLE void pauseProcess();
+  Q_INVOKABLE void resumeProcess();
 
 signals:
   void outputChanged();
   void isRunningChanged();
+  void isPausedChanged();
   void commandFinished(int exitCode, const QString &output);
 
 private slots:
@@ -42,10 +47,12 @@ private slots:
 private:
   void appendOutput(const QString &text);
   void setIsRunning(bool running);
+  void setIsPaused(bool paused);
 
   QProcess *m_process;
   QString m_output;
   bool m_isRunning;
+  bool m_isPaused;
 };
 
 #endif // PROCESSMANAGER_H